#977 Extract to util
